requirements_for_all_applications.txt
czifile==2019.7.2
ipython==8.13.2
numpy==1.23.4
Pillow==9.5.0
PyQt6==6.5.0
ultralytics==8.0.91
